chore: rename cron.ts to index.ts
