+++ examples/every_30_minutes_between_9_and_5.js
@@ -1,4 +1,4 @@
-const CronJob = require('../src/cron.js').CronJob;
+const CronJob = require('../dist').CronJob;
 
 console.log('Before job instantiation');
 const job = new CronJob('0 */30 9-17 * * *', function () {
