+++ examples/is_crontime_valid.js
@@ -1,4 +1,4 @@
-const CronJob = require('../src/cron.js').CronJob;
+const CronJob = require('../dist').CronJob;
 
 try {
 	new CronJob('NOT VALID', () => {
