+++ examples/mon_to_fri_at_11_30.js
@@ -1,4 +1,4 @@
-const CronJob = require('../src/cron.js').CronJob;
+const CronJob = require('../dist').CronJob;
 
 console.log('Before job instantiation');
 const job = new CronJob('00 30 11 * * 1-5', function () {
