+++ examples/object_param.js
@@ -1,17 +1,17 @@
-const CronJob = require('../src/cron.js').CronJob;
+const CronJob = require('../dist').CronJob;
 
 let isRunning = false;
 console.log('Before job instantiation');
 const job = new CronJob({
 	cronTime: '* * * * * *',
-	onTick: function() {
+	onTick: function () {
 		const d = new Date();
 		console.log('Check every second:', d, ', isRunning: ', isRunning);
 
 		if (!isRunning) {
 			isRunning = true;
 
-			setTimeout(function() {
+			setTimeout(function () {
 				console.log('Long running onTick complete:', new Date());
 				isRunning = false;
 			}, 3000);
